renamed allreduce
