--- include/core/allreduce_groups.h
@@ -1,16 +1,0 @@
-#ifndef EXT_MPI_ALLREDUCE_GROUPS_H_
-
-#define EXT_MPI_ALLREDUCE_GROUPS_H_
-
-#ifdef __cplusplus
-extern "C"
-{
-#endif
-
-int ext_mpi_generate_allreduce_groups(char *buffer_in, char *buffer_out);
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif
