+++ include/core/allreduce_single.h
@@ -1,0 +1,16 @@
+#ifndef EXT_MPI_ALLREDUCE_SINGLE_H_
+
+#define EXT_MPI_ALLREDUCE_SINGLE_H_
+
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+
+int ext_mpi_generate_allreduce_single(char *buffer_in, char *buffer_out);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
